Fix: Clean up code style and naming

+++ src/utils.h
@@ -1,4 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
-int factorial(int n);
+
+// Sửa từ 'int' thành 'unsigned long long' để khớp với file .c
+unsigned long long factorial(int n);
+
 #endif
